refactor 사용자 도메인 패키지 이동
