+++ hibenrnate-performance/src/main/java/com/gmoon/demo/member/repository/MemberRepositoryCustom.java
@@ -1,6 +1,6 @@
 package com.gmoon.demo.member.repository;
 
-import com.gmoon.demo.member.Member;
+import com.gmoon.demo.member.domain.Member;
 import com.gmoon.demo.member.model.Members;
 
 import java.util.List;
