+++ hibenrnate-performance/src/main/java/com/gmoon/demo/member/service/MemberService.java
@@ -1,6 +1,6 @@
 package com.gmoon.demo.member.service;
 
-import com.gmoon.demo.member.Member;
+import com.gmoon.demo.member.domain.Member;
 import com.gmoon.demo.member.model.MemberOptionUpdate;
 import com.gmoon.demo.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
