+++ hibenrnate-performance/src/test/java/com/gmoon/demo/member/repository/MemberRepositoryTest.java
@@ -3,8 +3,8 @@ package com.gmoon.demo.member.repository;
 import com.gmoon.demo.base.BaseRepositoryTest;
 import com.gmoon.demo.domain.QMember;
 import com.gmoon.demo.domain.QMemberOption;
-import com.gmoon.demo.member.Member;
-import com.gmoon.demo.member.MemberOption;
+import com.gmoon.demo.member.domain.Member;
+import com.gmoon.demo.member.domain.MemberOption;
 import lombok.RequiredArgsConstructor;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.DisplayName;
